* start authentication

* authentication start decorators

* login_required should work with access token

* backend authentication for most endpoints (very rough draft of functions in authentication.py)

* clean_up_function

* authentication cleanup

* give error when access_token fails

* documentation auth functions

* fixed imports

* actual import fix

* added requests

* authorize submissions

* removed double checks

* start testing setup backend authentication

* poging testen

* github tests check

* user tests with authentication

* auth url accessible hopefully

* change authorization to be easier to deal with since it doesn't matter for tests

* fixed jobCategory -> jobTitle

* fix authentication

* user tests zouden moeten slagen

* fix authentication arguments

* project tests with authentication

* changed auth server id of teacher

* maybe correct primary keys

* second try on primary key of course relations

* further test authentication

* authentication on project assignment files

* auth on course_join_codes and extra tests

* teacher_id in function when necessary

* user tests with authentication

* extra testing

* fixed comments

* lots of testing changes

* should be 1 error test now

* fix tests

+++ backend/tests/endpoints/project_test.py
@@ -11,11 +11,12 @@ def test_assignment_download(client, valid_project):
         response = client.post(
             "/projects",
             data=valid_project,
-            content_type='multipart/form-data'
+            content_type='multipart/form-data',
+            headers={"Authorization":"teacher2"}
         )
     assert response.status_code == 201
     project_id = response.json["data"]["project_id"]
-    response = client.get(f"/projects/{project_id}/assignments")
+    response = client.get(f"/projects/{project_id}/assignments", headers={"Authorization":"teacher2"})
     # file downloaded succesfully
     assert response.status_code == 200
 
@@ -26,19 +27,19 @@ def test_not_found_download(client):
     """
     response = client.get("/projects")
     # get an index that doesnt exist
-    response = client.get("/projects/-1/assignments")
+    response = client.get("/projects/-1/assignments", headers={"Authorization":"teacher2"})
     assert response.status_code == 404
 
 
 def test_projects_home(client):
     """Test home project endpoint."""
-    response = client.get("/projects")
+    response = client.get("/projects", headers={"Authorization":"teacher1"})
     assert response.status_code == 200
 
 
 def test_getting_all_projects(client):
     """Test getting all projects"""
-    response = client.get("/projects")
+    response = client.get("/projects", headers={"Authorization":"teacher1"})
     assert response.status_code == 200
     assert isinstance(response.json['data'], list)
 
@@ -52,14 +53,14 @@ def test_post_project(client, valid_project):
         response = client.post(
             "/projects",
             data=valid_project,
-            content_type='multipart/form-data'
+            content_type='multipart/form-data', headers={"Authorization":"teacher2"}
         )
 
     assert response.status_code == 201
 
     # check if the project with the id is present
     project_id = response.json["data"]["project_id"]
-    response = client.get(f"/projects/{project_id}")
+    response = client.get(f"/projects/{project_id}", headers={"Authorization":"teacher2"})
 
     assert response.status_code == 200
 
@@ -67,16 +68,16 @@ def test_remove_project(client, valid_project_entry):
     """Test removing a project to the datab and fetching it, testing if it's not present anymore"""
 
     project_id = valid_project_entry.project_id
-    response = client.delete(f"/projects/{project_id}")
+    response = client.delete(f"/projects/{project_id}", headers={"Authorization":"teacher2"})
     assert response.status_code == 200
 
     # check if the project isn't present anymore and the delete indeed went through
-    response = client.get(f"/projects/{project_id}")
+    response = client.get(f"/projects/{project_id}", headers={"Authorization":"teacher2"})
     assert response.status_code == 404
 
 def test_patch_project(client, valid_project_entry):
     """
-    Test functionality of the PUT method for projects
+    Test functionality of the PATCH method for projects
     """
 
     project_id = valid_project_entry.project_id
@@ -86,6 +87,6 @@ def test_patch_project(client, valid_project_entry):
 
     response = client.patch(f"/projects/{project_id}", json={
         "title": new_title, "archived": new_archived
-    })
+    }, headers={"Authorization":"teacher2"})
 
     assert response.status_code == 200
