+++ backend/tests/endpoints/submissions_test.py
@@ -14,186 +14,36 @@ class TestSubmissionsEndpoint:
     ### GET SUBMISSIONS ###
     def test_get_submissions_wrong_user(self, client: FlaskClient):
         """Test getting submissions for a non-existing user"""
-        response = client.get("/submissions?uid=-20")
+        response = client.get("/submissions?uid=-20", headers={"Authorization":"teacher1"})
         assert response.status_code == 400
 
     def test_get_submissions_wrong_project(self, client: FlaskClient):
         """Test getting submissions for a non-existing project"""
-        response = client.get("/submissions?project_id=-1")
-        assert response.status_code == 400
+        response = client.get("/submissions?project_id=-1", headers={"Authorization":"teacher1"})
+        assert response.status_code == 404 # can't find course of project in authorization
         assert "message" in response.json
 
     def test_get_submissions_wrong_project_type(self, client: FlaskClient):
         """Test getting submissions for a non-existing project of the wrong type"""
-        response = client.get("/submissions?project_id=zero")
+        response = client.get("/submissions?project_id=zero", headers={"Authorization":"teacher1"})
         assert response.status_code == 400
         assert "message" in response.json
 
-    def test_get_submissions_all(self, client: FlaskClient):
-        """Test getting the submissions"""
-        response = client.get("/submissions")
-        data = response.json
-        assert response.status_code == 200
-        assert "message" in data
-        assert isinstance(data["data"], list)
-
-    def test_get_submissions_user(self, client: FlaskClient, valid_submission_entry):
-        """Test getting the submissions given a specific user"""
-        response = client.get(f"/submissions?uid={valid_submission_entry.uid}")
-        data = response.json
-        assert response.status_code == 200
-        assert "message" in data
-
-
     def test_get_submissions_project(self, client: FlaskClient, valid_submission_entry):
         """Test getting the submissions given a specific project"""
-        response = client.get(f"/submissions?project_id={valid_submission_entry.project_id}")
+        response = client.get(f"/submissions?project_id={valid_submission_entry.project_id}", headers={"Authorization":"teacher2"})
         data = response.json
         assert response.status_code == 200
         assert "message" in data
 
-    def test_get_submissions_user_project(self, client: FlaskClient, valid_submission_entry):
-        """Test getting the submissions given a specific user and project"""
-        response = client.get(
-            f"/submissions? \
-                uid={valid_submission_entry.uid}&\
-                project_id={valid_submission_entry.project_id}")
-        data = response.json
-        assert response.status_code == 200
-        assert "message" in data
-
-    ### POST SUBMISSIONS ###
-    def test_post_submissions_no_user(self, client: FlaskClient, valid_project_entry, files):
-        """Test posting a submission without specifying a user"""
-        response = client.post("/submissions", data={
-            "project_id": valid_project_entry.project_id,
-            "files": files
-        })
-        data = response.json
-        assert response.status_code == 400
-        assert data["message"] == "The uid is missing"
-
-    def test_post_submissions_wrong_user(self, client: FlaskClient, valid_project_entry, files):
-        """Test posting a submission for a non-existing user"""
-        response = client.post("/submissions", data={
-            "uid": "unknown",
-            "project_id": valid_project_entry.project_id,
-            "files": files
-        })
-        data = response.json
-        assert response.status_code == 400
-        assert data["message"] == "Invalid user (uid=unknown)"
-
-    def test_post_submissions_no_project(self, client: FlaskClient, valid_user_entry, files):
-        """Test posting a submission without specifying a project"""
-        response = client.post("/submissions", data={
-            "uid": valid_user_entry.uid,
-            "files": files
-        })
-        data = response.json
-        assert response.status_code == 400
-        assert data["message"] == "The project_id is missing"
-
-    def test_post_submissions_wrong_project(self, client: FlaskClient, valid_user_entry, files):
-        """Test posting a submission for a non-existing project"""
-        response = client.post("/submissions", data={
-            "uid": valid_user_entry.uid,
-            "project_id": 0,
-            "files": files
-        })
-        data  = response.json
-        assert response.status_code == 400
-        assert data["message"] == "Invalid project (project_id=0)"
-
-    def test_post_submissions_wrong_project_type(
-            self, client: FlaskClient, valid_user_entry, files
-        ):
-        """Test posting a submission for a non-existing project of the wrong type"""
-        response = client.post("/submissions", data={
-            "uid": valid_user_entry.uid,
-            "project_id": "zero",
-            "files": files
-        })
-        data  = response.json
-        assert response.status_code == 400
-        assert data["message"] == "Invalid project_id typing (project_id=zero)"
-
-    def test_post_submissions_no_files(
-            self, client: FlaskClient, valid_user_entry, valid_project_entry):
-        """Test posting a submission when no files are uploaded"""
-        response = client.post("/submissions", data={
-            "uid": valid_user_entry.uid,
-            "project_id": valid_project_entry.project_id
-        })
-        data = response.json
-        assert response.status_code == 400
-        assert data["message"] == "No files were uploaded"
-
-
-    def test_post_submissions_empty_file(self, client: FlaskClient, session: Session, file_empty):
-        """Test posting a submission for an empty file"""
-        project = session.query(Project).filter_by(title="B+ Trees").first()
-        response = client.post("/submissions", data={
-            "uid": "student01",
-            "project_id": project.project_id,
-            "files": file_empty
-        })
-        data = response.json
-        assert response.status_code == 400
-        assert data["message"] == "No files were uploaded"
-
-    def test_post_submissions_file_with_no_name(
-            self, client: FlaskClient, session: Session, file_no_name
-        ):
-        """Test posting a submission for a file without a name"""
-        project = session.query(Project).filter_by(title="B+ Trees").first()
-        response = client.post("/submissions", data={
-            "uid": "student01",
-            "project_id": project.project_id,
-            "files": file_no_name
-        })
-        data = response.json
-        assert response.status_code == 400
-        assert data["message"] == "No files were uploaded"
-
-    def test_post_submissions_missing_required_files(
-            self, client: FlaskClient, session: Session, files
-        ):
-        """Test posting a submissions for a file with a wrong name"""
-        project = session.query(Project).filter_by(title="B+ Trees").first()
-        response = client.post("/submissions", data={
-            "uid": "student01",
-            "project_id": project.project_id,
-            "files": files
-        })
-        data = response.json
-        assert response.status_code == 400
-        assert data["message"] == "Not all required files were uploaded"
-
-    def test_post_submissions_correct(
-            self, client: FlaskClient, session: Session, files
-        ):
-        """Test posting a submission"""
-        project = session.query(Project).filter_by(title="Predicaten").first()
-        response = client.post("/submissions", data={
-            "uid": "student02",
-            "project_id": project.project_id,
-            "files": files
-        })
-        data = response.json
-        assert response.status_code == 201
-        assert data["message"] == "Successfully fetched the submissions"
-        assert data["url"] == f"{API_HOST}/submissions/{data['data']['id']}"
-        assert data["data"]["user"] == f"{API_HOST}/users/student02"
-        assert data["data"]["project"] == f"{API_HOST}/projects/{project.project_id}"
 
     ### GET SUBMISSION ###
     def test_get_submission_wrong_id(self, client: FlaskClient, session: Session):
         """Test getting a submission for a non-existing submission id"""
-        response = client.get("/submissions/0")
+        response = client.get("/submissions/0", headers={"Authorization":"ad3_teacher"})
         data = response.json
         assert response.status_code == 404
-        assert data["message"] == "Submission (submission_id=0) not found"
+        assert data["message"] == "Submission with id: 0 not found"
 
     def test_get_submission_correct(self, client: FlaskClient, session: Session):
         """Test getting a submission"""
@@ -201,7 +51,7 @@ class TestSubmissionsEndpoint:
         submission = session.query(Submission).filter_by(
             uid="student01", project_id=project.project_id
         ).first()
-        response = client.get(f"/submissions/{submission.submission_id}")
+        response = client.get(f"/submissions/{submission.submission_id}", headers={"Authorization":"ad3_teacher"})
         data = response.json
         assert response.status_code == 200
         assert data["message"] == "Successfully fetched the submission"
@@ -218,10 +68,10 @@ class TestSubmissionsEndpoint:
     ### PATCH SUBMISSION ###
     def test_patch_submission_wrong_id(self, client: FlaskClient, session: Session):
         """Test patching a submission for a non-existing submission id"""
-        response = client.patch("/submissions/0", data={"grading": 20})
+        response = client.patch("/submissions/0", data={"grading": 20}, headers={"Authorization":"ad3_teacher"})
         data = response.json
         assert response.status_code == 404
-        assert data["message"] == "Submission (submission_id=0) not found"
+        assert data["message"] == "Submission with id: 0 not found"
 
     def test_patch_submission_wrong_grading(self, client: FlaskClient, session: Session):
         """Test patching a submission with a wrong grading"""
@@ -229,7 +79,7 @@ class TestSubmissionsEndpoint:
         submission = session.query(Submission).filter_by(
             uid="student02", project_id=project.project_id
         ).first()
-        response = client.patch(f"/submissions/{submission.submission_id}", data={"grading": 100})
+        response = client.patch(f"/submissions/{submission.submission_id}", data={"grading": 100}, headers={"Authorization":"ad3_teacher"})
         data = response.json
         assert response.status_code == 400
         assert data["message"] == "Invalid grading (grading=0-20)"
@@ -240,18 +90,18 @@ class TestSubmissionsEndpoint:
         submission = session.query(Submission).filter_by(
             uid="student02", project_id=project.project_id
         ).first()
-        response = client.patch(f"/submissions/{submission.submission_id}",data={"grading": "zero"})
+        response = client.patch(f"/submissions/{submission.submission_id}",data={"grading": "zero"}, headers={"Authorization":"ad3_teacher"})
         data = response.json
         assert response.status_code == 400
         assert data["message"] == "Invalid grading (grading=0-20)"
 
-    def test_patch_submission_correct(self, client: FlaskClient, session: Session):
+    def test_patch_submission_correct_teacher(self, client: FlaskClient, session: Session):
         """Test patching a submission"""
         project = session.query(Project).filter_by(title="B+ Trees").first()
         submission = session.query(Submission).filter_by(
             uid="student02", project_id=project.project_id
         ).first()
-        response = client.patch(f"/submissions/{submission.submission_id}", data={"grading": 20})
+        response = client.patch(f"/submissions/{submission.submission_id}", data={"grading": 20}, headers={"Authorization":"ad3_teacher"})
         data = response.json
         assert response.status_code == 200
         assert data["message"] == f"Submission (submission_id={submission.submission_id}) patched"
@@ -265,14 +115,16 @@ class TestSubmissionsEndpoint:
             "path": "/submissions/2",
             "status": False
         }
+    
+    # TODO test course admin (allowed) and student (not allowed) patch 
 
     ### DELETE SUBMISSION ###
     def test_delete_submission_wrong_id(self, client: FlaskClient, session: Session):
         """Test deleting a submission for a non-existing submission id"""
-        response = client.delete("submissions/0")
+        response = client.delete("submissions/0", headers={"Authorization":"student01"})
         data = response.json
         assert response.status_code == 404
-        assert data["message"] == "Submission (submission_id=0) not found"
+        assert data["message"] == "Submission with id: 0 not found"
 
     def test_delete_submission_correct(self, client: FlaskClient, session: Session):
         """Test deleting a submission"""
@@ -280,7 +132,7 @@ class TestSubmissionsEndpoint:
         submission = session.query(Submission).filter_by(
             uid="student01", project_id=project.project_id
         ).first()
-        response = client.delete(f"submissions/{submission.submission_id}")
+        response = client.delete(f"submissions/{submission.submission_id}", headers={"Authorization":"student01"})
         data = response.json
         assert response.status_code == 200
         assert data["message"] == f"Submission (submission_id={submission.submission_id}) deleted"
